Add wxWebView
This patch adds support for the wxWebView component of wxWidgets. This finally enables Erlang developers to write modern desktop applications comparable in look and feel with state of the art applications.

The patchset ships with an addition of the demo application and adds a sample component in examples/demo/
This patchset also depends on adding libwxgtk-webveiw3.0-gtk3-dev to the base images for the CI builds to succeed.